+++ src/ParlorPrediction.Mvc/Models/DoughUsage/DoughUsageTraceListItemViewModel.cs
@@ -12,7 +12,7 @@ public sealed class DoughUsageTraceListItemViewModel
 
     public string Destination { get; set; } = string.Empty;
 
-    public int TrayCount { get; set; }
+    public decimal TrayCount { get; set; }
 
     public int BallsUsed { get; set; }
